+++ src/mozilla_sec_eia/library/validation_helpers.py
@@ -6,9 +6,6 @@ from importlib import resources
 
 import pandas as pd
 
-from mozilla_sec_eia.models.sec10k.ex_21.inference import clean_extracted_df
-from mozilla_sec_eia.models.sec10k.utils.cloud import get_metadata_filename
-
 
 def load_validation_data(
     filename: str, index_cols: list[str] | None = None
@@ -75,22 +72,6 @@ def jaccard_similarity(
     )
     union = set(computed_df[value_col]).union(set(validation_df[value_col]))
     return float(len(intersection)) / float(len(union))
-
-
-def clean_ex21_validation_set(validation_df: pd.DataFrame):
-    """Clean Ex. 21 validation data to match extracted format."""
-    validation_df = validation_df.rename(
-        columns={
-            "Filename": "id",
-            "Subsidiary": "subsidiary",
-            "Location of Incorporation": "loc",
-            "Ownership Percentage": "own_per",
-        }
-    )
-    validation_df["own_per"] = validation_df["own_per"].astype(str)
-    validation_df["filename"] = validation_df["id"].apply(get_metadata_filename)
-    validation_df = clean_extracted_df(validation_df)
-    return validation_df
 
 
 def strip_down_company_names(ser: pd.Series) -> pd.Series:
