+++ src/mozilla_sec_eia/models/sec10k/ex_21/validation_helpers.py
@@ -1,0 +1,22 @@
+"""Helper functions specific to Exhibit 21 model validation."""
+
+import pandas as pd
+
+from mozilla_sec_eia.models.sec10k.ex_21.inference import clean_extracted_df
+from mozilla_sec_eia.models.sec10k.utils.cloud import get_metadata_filename
+
+
+def clean_ex21_validation_set(validation_df: pd.DataFrame):
+    """Clean Ex. 21 validation data to match extracted format."""
+    validation_df = validation_df.rename(
+        columns={
+            "Filename": "id",
+            "Subsidiary": "subsidiary",
+            "Location of Incorporation": "loc",
+            "Ownership Percentage": "own_per",
+        }
+    )
+    validation_df["own_per"] = validation_df["own_per"].astype(str)
+    validation_df["filename"] = validation_df["id"].apply(get_metadata_filename)
+    validation_df = clean_extracted_df(validation_df)
+    return validation_df
